feat: notify 전역관리
